version 0.4 Добавил: базовую логику для OrderService и OrderProductService.

+++ src/main/java/com/pavcore/plumbing/repo/OrderRepo.java
@@ -3,5 +3,5 @@ package com.pavcore.plumbing.repo;
 import com.pavcore.plumbing.entiy.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepo extends JpaRepository<Order, Integer> {
+public interface OrderRepo extends JpaRepository<Order, Long> {
 }
